+++ changelog.md
@@ -1,6 +1,7 @@
 # YouCanDoIt Changelog
 ## v0.0.1-alpha.4
 - Fix even more residual bugs.
+- Add easing types.
 ## v0.0.1-alpha.3
 - Fix some residual bugs.
 ## v0.0.1-alpha.2